Correção em tipo de Profile, aumentando tamanho do nome do curso

+++ planner/src/main/java/edu/planner/repositories/IUserRepo.java
@@ -11,9 +11,9 @@ public interface IUserRepo extends PagingAndSortingRepository<User, Long> {
 
 	public Page<User> findByNameContaining(Pageable page, String description);
 
-	public Page<User> findByProfilesInAndNameContaining(Pageable page, Integer profile, String description);
+	public Page<User> findByProfilesInAndNameContaining(Pageable page, Short profile, String description);
 	
-	public Page<User> findByProfilesIs(Pageable page, Integer profile);
+	public Page<User> findByProfilesIs(Pageable page, Short profile);
 
 	@Transactional(readOnly=true)
 	public User findByEmail(String email);
